requirements.txt
Django==4.2
django-jazzmin==2.6.0
django-jsonform==2.17.0
django-import-export==3.2.0
